Add a queueing mechanism to pass results from workers to the calling method.
A worker can put results in the queue as soon as it has completed one job. It
means that the calling method has access to the results earlier than in
previous implementation and can start processing results on the fly instead of
remaining idle while waiting for all jobs to complete.